+++ packages/pkg-core/src/schematics/schematics-resettable-node-workflow.ts
@@ -1,11 +1,3 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
 import { Sink, UnsuccessfulWorkflowExecution } from "@angular-devkit/schematics";
 import { NodeWorkflow } from "@angular-devkit/schematics/tools";
 import { Observable, of, Subject, throwError } from "rxjs";