+++ packages/pkg-core/src/schematics/tasks/service-exec/executor.ts
@@ -1,11 +1,3 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
 import { SchematicContext, TaskExecutor } from '@angular-devkit/schematics';
 import { assertUtils } from '../../../utils';
 import { Services } from '../../../__types__';