+++ packages/plugin-core-essentials/src/schematics/workspace-init/factory.ts
@@ -1,11 +1,3 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
 import { strings } from "@angular-devkit/core";
 import {
   apply,
@@ -19,7 +11,6 @@ import {
 } from "@angular-devkit/schematics";
 import { NodePackageInstallTask } from "@angular-devkit/schematics/tasks";
 import { expectIsOk, schematicUtils } from "@business-as-code/core";
-import { schematicTestUtils } from "@business-as-code/tests-core";
 import { Schema } from "./schema";
 
 export default function (options: Schema): Rule {
@@ -43,7 +34,7 @@ export default function (options: Schema): Rule {
       // move(destinationPath),
     ]);
 
-    const pmTaskHandle = schematicContext.addTask(new NodePackageInstallTask({workingDirectory: '.', quiet: false, hideOutput: false, packageManager: 'pnpm'}), []);
+    const pmTaskHandle = schematicContext.addTask(new NodePackageInstallTask({workingDirectory: '.',  quiet: false, hideOutput: false, packageManager: 'pnpm'}), []);
     // const pmTaskHandle2 = schematicContext.addTask(new NodePackageInstallTask({workingDirectory: '.', quiet: false, hideOutput: false, packageManager: 'pnpm'}), [pmTaskHandle]);
     // const addDepTaskHandle = schematicContext.addTask(new NodePackageInstallTask({packageName: '@business-as-code/cli', workingDirectory: '.', quiet: false, hideOutput: false, packageManager: 'pnpm'}), [pmTaskHandle]);
 
@@ -100,7 +91,7 @@ export default function (options: Schema): Rule {
       //   },
       //   schematicContext,
       // }),
-      schematicTestUtils.debugRule({context: options._bacContext,
+      schematicUtils.debugRule({context: options._bacContext,
         initialiseOptions: {
           workingPath: '.',
         },}),