+++ packages/plugin-core-git/src/schematics/git-clone/factory.ts
@@ -1,11 +1,3 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
 import { chain, empty, mergeWith, move, Rule } from "@angular-devkit/schematics";
 import { schematicUtils } from "@business-as-code/core";
 import { Schema } from "./schema";