+++ packages/tests/pkg-tests-core/src/schematics/index.ts
@@ -1,1 +1,2 @@
-export * as schematicTestUtils from './schematics-utils'
+// export * as schematicTestUtils from './schematics-utils'
+export {}