+++ packages/pkg-fslib/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @business-as-code/fslib
 
+## 0.0.0-bollards-20230521070624
+
+### Patch Changes
+
+- new message
+- new message
+
+## 0.0.0-bollards-20230521065148
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520194736
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520191024
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520190936
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520190630
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520184509
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520184209
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520183922
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520182953
+
+### Patch Changes
+
+- new message
+
 ## 0.0.0-bollards-20230520182205
 
 ### Patch Changes